pyproject.toml
[project]
name = "cian"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "playwright>=1.58.0",
    "sqlalchemy>=2.0",
    "psycopg[binary]>=3.2",
    "httpx>=0.27",
    "pillow>=10.4",
    "python-dotenv>=1.0",
    "flask>=3.0",
]
